+++ metadata.rb
@@ -1,9 +1,9 @@
-name             'r-nativex'
-maintainer       'NativeX'
-maintainer_email 'adrian.herrera@nativex.com'
+name             'r-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures r-nativex'
+description      'Installs/Configures r-simplyadrian'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.2'
 
-depends		 "r"
+depends  'r'