+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-r-nativex CHANGELOG
+r-simplyadrian CHANGELOG
 ===================
 
-This file is used to list changes made in each version of the r-nativex cookbook.
+This file is used to list changes made in each version of the r-simplyadrian cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of r-nativex
+- [Adrian Herrera] - Initial release of r-simplyadrian
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.